pricing plan steps are complete
Pricing plan feature fails (of course) because expected price of $49
doesn't equal actual price on website of $99.

Next: need to do guinea_pig.feature (and steps)

But it should be easier now that I've gotten the support stuff sorted
out.

Then, I still need to set it to use sauce labs for configuration.

I'm still not sure how to do parallelization because Javascript is
single-threaded and I haven't found an example to parallelize.